Resolve interaction between BB8 and Push the Limit.
Previously PtL was marked as used after the first time it was offered to
the user. A free action from PtL is now only offered following
a successfully completed action and if PtL has not already been used.

This is not the best possible solution. Upcoming refactoring should allow
for the result of an action (taken/skipped) to be observed thus reducing
the logic for determining if PtL can be taken. This refactoring should
also remove the possibility of an infinite loop occuring do to the fact
that PtL occurs following an action and itself gives and completes an
action. Thus calling itself following the resolution of itself.