Fix for barrel roll on second edition
For some reason on mobile, when the UpdateColisionDetection function returns true, ShipMovementScript.UpdateSubscribedFuncs() doesn't actually remove theUpdateColisionDetection function from its subscription list. It calls FuncsToUpdate.Remove() on the function, but after that the function is still in the list somehow.

So to fix that we simply return true again the next time UpdateColisionDetection is called, and that time it actually gets removed.