Focus, Reload and Calculate
Now boost Focus up to Target Lock level if the ship has a weapon that requires focus tokens.

Also boosts Calculate up to Target Lock level if the ship has a weapon that requires Calculate and it has charges.

Reload priority test is now based on whether this is First or Second edition.