Azrapse delayed fuses (#1939)
* Added Delayed Fuses
Bomb and Mine prefabs are now of the new GenericDeviceGameObject type (instead of just GameObjects).
GenericDeviceGameObject has Fuses and IsFused properties that keep tracks of the amount of fuse markers on that device object.
On creation, it spawns a DeviceInfoPanel child canvas object that displays how many fuse tokens there are present on the device, if at least 1.
The DeviceInfoPanel displays an animated icon based on the actual fuse marker icon by FFG, and a LCD-like display indicating how many markers there are in the form of "T-2" if there are 2 markers, for example.
Added Fuse rule, that is called when determining permission to detonate. This rule decrements the amount of fuse markers on the current detonating device and prevents detonation.

Note: The change of type from GameObject to GenericDeviceGameObject for bombs and mines causes many cascaded type castings and variable definitions in multiple files. That is why this commit changes so many files.

* Electro-Proton Bomb
Added a new field to bombs: detonationRange. Defaults to 1 for all GenericTimed bombs, but it's 2 for EPB.

* Strangely, EPB's material didn't have any albedo on project reopen... fixed.

* Cost updated for Electro-Proton Bomb and Delayed Fuses to match PDF.